requirements.txt
numpy
pyo
asyncio
matplotlib
websockets
requests
python-dotenv
